[REFACTOR] utils에서 파일 이동
