--- src/utils/env.ts
@@ -1,3 +1,0 @@
-export const isLocal = process.env.REACT_APP_MODE === 'local';
-export const isDevelopment = process.env.REACT_APP_MODE === 'development';
-export const isProduction = process.env.REACT_APP_MODE === 'production';
