--- src/utils/helpers.ts
@@ -1,3 +1,0 @@
-export function colorSelector(key, token) {
-    return token.find((t) => t.description === key).value;
-}
